Allow finding all nodes and nodes containing a string (#545)
When using the `-f` option in `rust-code-analysis-cli` to find nodes, the current behavior is to return all found nodes if the passed filter string is not one of `call|comment|error|string|function|<u16>`. I found it useful to be able to find nodes that contained a certain string (e.g. `expr`) so I added this functionality. Because in testing I found that an empty `-f` never creates the "accept all" filter, I added a keyword `"all"` to enable this behavior (also quite useful). The filters now accept `all|call|comment|error|string|function|<u16>|<string>`.